moved download handlers to their own module. #136.
